feature: Session 성공 테스트 작성 완료

+++ src/main/java/nextstep/courses/domain/port/SessionCoverImageRepository.java
@@ -6,5 +6,5 @@ public interface SessionCoverImageRepository {
 
     SessionCoverImage findById(Long id);
 
-    int save(SessionCoverImage sessionCoverImage);
+    Long save(SessionCoverImage sessionCoverImage);
 }
